# contracts/openapi/incident-api.yaml
openapi: 3.0.3
info:
  title: Telecom Service Assurance API
  version: 0.3.0
  description: Revision 3 contracts. Service KPI trends and incident episodes replace subscriber risk scoring.
    Application endpoints remain implementation tasks; this document is not evidence of a running API.
servers:
- url: http://telecom.test:8080
  description: Canonical browser origin through NGINX; incident-service uses internal port 8082.
security:
- sessionCookie: []
...
paths:
...
  /api/incidents:
    get:
      operationId: listIncidents
      summary: List incident episodes
      tags:
      - Incidents
      description: Sort detectedAt DESC,id DESC. One row per episode, not per breached minute.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/Unavailable'
      parameters:
      - name: service
        in: query
        required: false
        schema:
          type: string
          enum:
          - VOLTE
          - SMS
      - name: scopeId
        in: query
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_.:-]{1,96}$
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - OPEN
          - INVESTIGATING
          - RESOLVED
      - name: technicalState
        in: query
        required: false
        schema:
          type: string
          enum:
          - ONGOING
          - RECOVERED
          - UNKNOWN
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
  /api/incidents/{id}:
    get:
      operationId: getIncident
      summary: Get incident episode
      tags:
      - Incidents
      description: Return one analyst work item and its latest accepted detection. Historical detections are
        available from the separate paged detections endpoint.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/Unavailable'
      parameters:
      - $ref: '#/components/parameters/IncidentId'
...
  /api/incidents/{id}/status:
    post:
      operationId: changeIncidentStatus
      summary: Change analyst workflow status
      tags:
      - Incidents
      description: Assigned analyst or supervisor/admin. OPEN -> INVESTIGATING requires assignee. INVESTIGATING
        -> RESOLVED requires technicalState RECOVERED and nonblank resolutionNote. Version conflict returns
        409. Technical recovery never resolves analyst work automatically.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/Unavailable'
      parameters:
      - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeStatusRequest'
      security:
      - sessionCookie: []
        csrfToken: []
  /api/incidents/{id}/assignment:
    post:
      operationId: assignIncident
      summary: Assign an incident
      tags:
      - Incidents
      description: Analyst can claim unassigned work for self; supervisors/admins can reassign to enabled analysts.
        Optimistic locking and audit are transactional.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/Unavailable'
      parameters:
      - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentRequest'
      security:
      - sessionCookie: []
        csrfToken: []
  /api/incidents/{id}/comments:
    post:
      operationId: commentOnIncident
      summary: Add a comment
      tags:
      - Incidents
      description: Assigned analyst or supervisor/admin; requestId provides comment idempotency.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/Unavailable'
      parameters:
      - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
      security:
      - sessionCookie: []
        csrfToken: []
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Opaque server-session identifier, not a JWT. HttpOnly, SameSite=Lax, Path=/; Secure with
        HTTPS. 15-minute idle and 30-minute absolute expiry.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: Token returned by GET /api/auth/csrf. Required alongside the session cookie for JSON mutations;
        never a replacement for authentication. POST /logout uses the _csrf form parameter instead.
  parameters:
...
    IncidentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Incident UUID, not a detectionId.
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: Malformed input, an invalid range or invalid request identifier.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthenticated:
      description: No valid server session, or the session has expired. API requests receive JSON, not a login
        redirect.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Insufficient application role, disabled local analyst, wrong assignee or failed CSRF validation.
        CSRF rejection uses code CSRF_INVALID.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Stale version or an incompatible workflow/assignment state. State failures use INVALID_TRANSITION
        or ASSIGNMENT_CONFLICT; currentVersion may be returned.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unavailable:
      description: A required dependency is unavailable; no secrets or internal exception text is returned.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    IncidentLocation:
      type: object
      additionalProperties: false
      description: Resolved from the opening detection's captured topology and its exact source-window coverage catalogue when unique; otherwise requires a unique historical topology binding. Multiple matching coverage catalogues remain ambiguous. Today's mapping never replaces opening evidence, and legacy scopes stay unallocated.
      required: [cityId, catalogueVersion, topologyVersion, measuredScopeId, containmentPath, dependencyNodeIds, nullReason]
      properties:
        cityId: {type: string, nullable: true}
        catalogueVersion: {type: string, nullable: true}
        topologyVersion: {type: string, nullable: true}
        measuredScopeId: {type: string}
        containmentPath:
          type: array
          maxItems: 5
          items: {type: string}
        dependencyNodeIds:
          type: array
          items: {type: string}
        nullReason:
          type: string
          nullable: true
          enum: [UNALLOCATED, CAPTURED_VERSION_MISSING, BINDING_NOT_FOUND, AMBIGUOUS_CATALOGUE_VERSION, INVALID_CONTAINMENT]
...
    Sha256Id:
      type: string
      pattern: ^[a-f0-9]{64}$
      description: Lowercase SHA-256 hexadecimal identity derived from the canonical episode contract.
...
    ApiError:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      - requestId
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
          format: uuid
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
            - field
            - message
            properties:
              field:
                type: string
              message:
                type: string
        currentVersion:
          type: integer
          minimum: 0
          format: int64
...
    ChangeStatusRequest:
      type: object
      additionalProperties: false
      required:
      - status
      - version
      properties:
        status:
          type: string
          enum:
          - INVESTIGATING
          - RESOLVED
        version:
          type: integer
          minimum: 0
          format: int64
        resolutionNote:
          type: string
          minLength: 1
          maxLength: 2000
          pattern: \S
      description: Service requires a note for RESOLVED, rejects direct OPEN -> RESOLVED, and requires assignment
        before INVESTIGATING. The acting analyst must be the assignee unless SUPERVISOR/ADMIN.
    AssignmentRequest:
      type: object
      additionalProperties: false
      required:
      - analystId
      - version
      properties:
        analystId:
          type: string
          format: uuid
        version:
          type: integer
          minimum: 0
          format: int64
      description: Assign to an enabled local analyst. ANALYST may only self-claim an unassigned incident;
        SUPERVISOR/ADMIN may reassign. A concurrent claim is a 409.
    CommentRequest:
      type: object
      additionalProperties: false
      required:
      - text
      - version
      - requestId
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 2000
          pattern: \S
        version:
          type: integer
          minimum: 0
          format: int64
        requestId:
          type: string
          format: uuid
      description: requestId is correlation metadata, never actor identity. If X-Request-ID is also present,
        it must match this value.
    ServiceDetection:
      type: object
      additionalProperties: false
      properties:
        schemaVersion:
          type: integer
          enum:
          - 2
        detectionId:
          $ref: '#/components/schemas/Sha256Id'
        episodeId:
          $ref: '#/components/schemas/Sha256Id'
        correlationKey:
          $ref: '#/components/schemas/Sha256Id'
        sequence:
          type: integer
          minimum: 1
        phase:
          type: string
          enum:
          - OPEN
          - UPDATE
          - RECOVERY
          - UNKNOWN
        anomalyType:
          type: string
          enum:
          - VOLTE_SETUP_DEGRADATION
          - SMS_DELIVERY_DELAY
        scopeId:
          type: string
          pattern: ^[A-Za-z0-9_.:-]{1,96}$
        service:
          type: string
          enum:
          - VOLTE
          - SMS
        firstObservedAt:
          type: string
          format: date-time
          pattern: Z$
          description: Episode anchor and first eligible breached window start used to derive episodeId.
        windowStart:
          type: string
          format: date-time
          pattern: Z$
        windowEnd:
          type: string
          format: date-time
          pattern: Z$
        detectedAt:
          type: string
          format: date-time
          pattern: Z$
        severity:
          type: string
          enum:
          - MEDIUM
          - HIGH
          - CRITICAL
        technicalState:
          type: string
          enum:
          - ONGOING
          - RECOVERED
          - UNKNOWN
        rulesetVersion:
          type: string
          minLength: 1
          maxLength: 160
        baselineVersion:
          type: string
          minLength: 1
          maxLength: 160
        topologyVersion:
          type: string
          minLength: 1
          maxLength: 160
        kpis:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              name:
                type: string
                minLength: 1
                maxLength: 160
              observed:
                type: number
                nullable: true
              baseline:
                type: number
                nullable: true
              unit:
                type: string
                enum:
                - PERCENT
                - PERCENTAGE_POINTS
                - RATIO
                - COUNT
                - MILLISECONDS
                - SECONDS
                - MBPS
              numerator:
                type: number
                nullable: true
              denominator:
                type: number
                nullable: true
            required:
            - name
            - observed
            - baseline
            - unit
            - numerator
            - denominator
          maxItems: 30
        impact:
          type: object
          additionalProperties: false
          properties:
            extraFailedAttempts:
              type: number
              minimum: 0
            affectedDeliveredMessages:
              type: integer
              minimum: 0
            pendingMessages:
              type: integer
              minimum: 0
            uniqueSubscribers:
              type: string
              nullable: true
              enum:
              - null
              description: Always null in this aggregate demo; observations do not identify unique subscribers.
          required:
          - extraFailedAttempts
          - affectedDeliveredMessages
          - pendingMessages
          - uniqueSubscribers
        probableCause:
          type: string
          maxLength: 1000
        causeConfidence:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
        evidence:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              code:
                type: string
                minLength: 1
                maxLength: 160
              summary:
                type: string
                maxLength: 1000
              nodeId:
                type: string
                nullable: true
              sourceEventIds:
                type: array
                items:
                  type: string
                  format: uuid
                maxItems: 20
            required:
            - code
            - summary
            - nodeId
            - sourceEventIds
          maxItems: 20
        recommendedChecks:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 160
          maxItems: 10
        mlStatus:
          type: string
          enum:
          - OK
          - TIMEOUT
          - UNAVAILABLE
          - INSUFFICIENT_DATA
          - NOT_APPLICABLE
        modelVersion:
          type: string
          nullable: true
        anomalyRank:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
      required:
      - schemaVersion
      - detectionId
      - episodeId
      - correlationKey
      - sequence
      - phase
      - anomalyType
      - scopeId
      - service
      - firstObservedAt
      - windowStart
      - windowEnd
      - detectedAt
      - severity
      - technicalState
      - rulesetVersion
      - baselineVersion
      - topologyVersion
      - kpis
      - impact
      - probableCause
      - causeConfidence
      - evidence
      - recommendedChecks
      - mlStatus
      - modelVersion
      - anomalyRank
...
    Incident:
      type: object
      additionalProperties: false
      properties:
        id: &id004
          type: string
          format: uuid
        episodeId:
          $ref: '#/components/schemas/Sha256Id'
        service:
          type: string
          enum:
          - VOLTE
          - SMS
        scopeId: &id005
          type: string
          pattern: ^[A-Za-z0-9_.:-]{1,96}$
        status:
          type: string
          enum:
          - OPEN
          - INVESTIGATING
          - RESOLVED
        technicalState:
          type: string
          enum:
          - ONGOING
          - RECOVERED
          - UNKNOWN
        severity:
          type: string
          enum:
          - MEDIUM
          - HIGH
          - CRITICAL
        assigneeId:
          type: string
          format: uuid
          nullable: true
        resolutionNote:
          type: string
          nullable: true
        firstObservedAt: &id003
          type: string
          format: date-time
          pattern: Z$
        detectedAt: *id003
        lastObservedAt: *id003
        createdAt: *id003
        updatedAt: *id003
        version:
          type: integer
          minimum: 0
        latestSequence:
          type: integer
          minimum: 1
        latestDetection:
          $ref: '#/components/schemas/ServiceDetection'
        location:
          $ref: '#/components/schemas/IncidentLocation'
        presentation:
          type: object
          additionalProperties: false
          description: Impact freshness and explanation copied from latest immutable detection.
          properties:
            impactState:
              type: string
              enum: [CURRENT, STALE, RECOVERED]
            currentImpact:
              type: object
              nullable: true
              additionalProperties: false
              properties:
                extraFailedAttempts: {type: number, minimum: 0}
                affectedDeliveredMessages: {type: integer, minimum: 0}
                pendingMessages: {type: integer, minimum: 0}
                uniqueSubscribers: {type: string, nullable: true, enum: [null]}
              required: [extraFailedAttempts, affectedDeliveredMessages, pendingMessages, uniqueSubscribers]
            retainedImpact:
              type: object
              nullable: true
              additionalProperties: false
              description: Last observed estimate during UNKNOWN, never a live estimate or historical peak.
              properties:
                extraFailedAttempts: {type: number, minimum: 0}
                affectedDeliveredMessages: {type: integer, minimum: 0}
                pendingMessages: {type: integer, minimum: 0}
                uniqueSubscribers: {type: string, nullable: true, enum: [null]}
              required: [extraFailedAttempts, affectedDeliveredMessages, pendingMessages, uniqueSubscribers]
            impactSourceDetectionId:
              type: string
            impactWindowEnd:
              type: string
              format: date-time
            probableCause:
              type: string
            causeConfidence:
              type: string
              enum: [LOW, MEDIUM, HIGH]
            evidenceHistoryPath:
              type: string
          required: [impactState, currentImpact, retainedImpact, impactSourceDetectionId,
            impactWindowEnd, probableCause, causeConfidence, evidenceHistoryPath]
      required:
      - id
      - episodeId
      - service
      - scopeId
      - status
      - technicalState
      - severity
      - assigneeId
      - resolutionNote
      - firstObservedAt
      - detectedAt
      - lastObservedAt
      - createdAt
      - updatedAt
      - version
      - latestSequence
      - latestDetection
      - presentation
    IncidentPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Incident'
          maxItems: 100
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
      required:
      - items
      - total
      - page
      - size
...